==> Code/16_make_summary_tables.Rmd <==
## Purpose: Summarize fire deverity in vegetation, ecoregion, fire event categories, TSLF categories,
## Project: 2020_Fires_GEB
## Upstream: 15_thin_grid.R
## Downstream: (none)


---
title: "2020 CA fire summary stats"
output:
  html_notebook:
    code_folding: hide
  html_document:
    df_print: paged
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
library(sf)
library(tidyverse)
library(here)
library(knitr)
library(DT)

data_dir = readLines(here("data_dir.txt"), n=1) # The location of root of the data directory on your computer should be specified in "data_dir.txt" (ignored by git)
source(here("Code/00_shared_functions_and_globals.R")) # This defines the function `datadir()` which makes it easy to refer to data that is stored outside the repo

d = st_read(datadir("CleanData/grid_thinned_02.gpkg"))
d_backup = d
st_geometry(d) = NULL

# load ecoregion
er_crosswalk = read_csv(datadir("AncillaryData/EcologicalSection_crosswalk.csv"))
d = d %>%
  left_join(er_crosswalk,by=c(ecrgn_s="code"))

# break TSLF into categories
d = d %>%
  mutate(TSLF_cat = cut(TSLF,breaks=c(-1,0,5,15,30,50,1000),labels=c("0","1-5","6-15","16-30","31-50","50+"), include.lowest=TRUE))


```


## Summary tables for percent high severity (PHS) for different classes
This is using the 2x2 thinned dataset (points every 180 m NS and EW)

### Percent high severity by veg. type
Only including CWHR types with > 5000 points

```{r}
phs_by_cwhr = d %>%
  group_by(cwhr_gp) %>%
  summarize(phs = sum(rdnbr_h)/n(),
            n_pts = n()) %>%
  mutate(phs = round(phs,2)) %>%
  filter(n_pts > 5000)

sig_cwhr = unique(phs_by_cwhr$cwhr_gp)

datatable(phs_by_cwhr,options=list(pageLength=100, lengthChange=FALSE, dom='t', columnDefs =          list(list(className = 'dt-center', targets = "_all"))), width=500)
```
### Percent high severity by ecoregion
Only including ecoregions with > 5000 points
```{r}
phs_by_ecoregion = d %>%
  group_by(ecoregion_name) %>%
  summarize(phs = sum(rdnbr_h)/n(),
            n_pts = n()) %>%
  mutate(phs = round(phs,2)) %>%
  filter(n_pts > 5000)

sig_ecoregion = unique(phs_by_ecoregion$ecoregion_name)

datatable(phs_by_ecoregion,options=list(pageLength=100, lengthChange=FALSE, dom='t', columnDefs =          list(list(className = 'dt-center', targets = "_all"))), width=500)
```
### Percent high severity by CWHR and ecoregion
Only including CWHR-ecoregion combinations with > 1000 points
```{r}
phs_by_cwhr_and_ecoregion = d %>%
  filter(cwhr_gp %in% sig_cwhr,
         ecoregion_name %in% sig_ecoregion) %>%
  group_by(cwhr_gp,ecoregion_name) %>%
  summarize(phs = sum(rdnbr_h)/n(),
            n_pts = n()) %>%
  mutate(phs = round(phs,2)) %>%
  filter(n_pts > 1000) %>%
  select(-n_pts)

# make it wide
phs_wide = phs_by_cwhr_and_ecoregion %>%
  mutate(across(everything(),as.character)) %>%
  pivot_wider(names_from=ecoregion_name,values_from=phs,values_fill = list(phs="--"))
  
## bring in an overall (across all ecoregions) PHS
phs_by_cwhr2 = phs_by_cwhr %>%
  select(cwhr_gp,OVERALL = phs)
phs_wide = phs_wide %>%
  left_join(phs_by_cwhr2)

## bring in an overall (across all CWHR types) PHS, as a row
phs_by_ecoregion2 = phs_by_ecoregion
phs_by_ecoregion2 = phs_by_ecoregion2 %>%
  as.data.frame() %>%
  select(ecoregion_name,OVERALL = phs) %>%
  t()

colnames(phs_by_ecoregion2) = phs_by_ecoregion2["ecoregion_name",]
phs_by_ecoregion2 = phs_by_ecoregion2["OVERALL",]

phs_wide = phs_wide %>%
  bind_rows(phs_by_ecoregion2) %>%
  mutate(cwhr_gp = as.character(cwhr_gp))

phs_wide[nrow(phs_wide),]$cwhr_gp = "OVERALL"

datatable(phs_wide,options=list(pageLength=500, lengthChange=FALSE, dom='t', columnDefs =          list(list(className = 'dt-center', targets = "_all"))), width=500)
```

### Percent high severity by TSLF
```{r}
phs_by_tslf = d %>%
  group_by(TSLF_cat) %>%
  summarize(phs = sum(rdnbr_h)/n(),
            n_pts = n()) %>%
  mutate(phs = round(phs,2))

sig_tslf = unique(phs_by_tslf$TSLF_cat)

datatable(phs_by_tslf,options=list(pageLength=100, lengthChange=FALSE, dom='t', columnDefs =          list(list(className = 'dt-center', targets = "_all"))), width=500)
```

### Percent high severity by CWHR and TSLF
Only including CWHR-TSLF combinations with > 1000 points
```{r}
phs_by_cwhr_and_tslf = d %>%
  filter(cwhr_gp %in% sig_cwhr) %>%
  group_by(cwhr_gp,TSLF_cat) %>%
  summarize(phs = sum(rdnbr_h)/n(),
            n_pts = n()) %>%
  mutate(phs = round(phs,2)) %>%
  filter(n_pts > 1000) %>%
  select(-n_pts)

# make it wide
phs_wide = phs_by_cwhr_and_tslf %>%
  mutate(across(everything(),as.character)) %>%
  pivot_wider(names_from=TSLF_cat,values_from=phs,values_fill = list(phs="--"))

datatable(phs_wide,options=list(pageLength=500, lengthChange=FALSE, dom='t', columnDefs =          list(list(className = 'dt-center', targets = "_all"))), width=500)
```


### Percent high severity by Fire
Including only fires with > 500 points
```{r}
phs_by_fire = d %>%
  group_by(FIRE_NA) %>%
  summarize(phs = sum(rdnbr_h)/n(),
            n_pts = n()) %>%
  mutate(phs = round(phs,2)) %>%
  filter(n_pts > 500)

sig_fire = unique(phs_by_fire$FIRE_NA)

datatable(phs_by_fire,options=list(pageLength=100, lengthChange=FALSE, dom='t', columnDefs =          list(list(className = 'dt-center', targets = "_all"))), width=500)
```
